@@ -43,29 +54,36 @@
     "jwt-decode": "^3.1.2",
     "million": "^2.6.4",
     "nanoid": "^3.3.4",
-    "next": "14.0.3",
+    "next": "^14.2.4",
     "next-pwa": "^5.6.0",
+    "patch-package": "^8.0.0",
     "posthog-js": "^1.103.1",
+    "posthog-node": "^4.0.1",
     "prosemirror-model": "1.21.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.43.9",
+    "react-infinite-scroller": "^1.2.6",
     "react-masonry-css": "^1.0.16",
     "react-qr-code": "^2.0.12",
+    "redoc": "^2.1.5",
     "sharp": "^0.30.7",
+    "svix": "^1.24.0",
+    "swagger-ui": "^5.17.14",
+    "swagger-ui-react": "^5.17.14",
     "tabler-icons-react": "^1.56.0"
   },
   "devDependencies": {
     "@babel/core": "^7.17.8",
     "@next/bundle-analyzer": "^14.0.3",
     "@next/eslint-plugin-next": "^12.1.4",
-    "@storybook/addon-essentials": "^7.6.7",
-    "@storybook/addon-interactions": "^7.6.7",
-    "@storybook/addon-links": "^7.6.7",
-    "@storybook/addon-onboarding": "^1.0.10",
-    "@storybook/addon-styling-webpack": "^0.0.6",
-    "@storybook/nextjs": "^7.6.7",
-    "@storybook/react": "^7.6.7",
+    "@storybook/addon-essentials": "^8.2.1",
+    "@storybook/addon-interactions": "^8.2.1",
+    "@storybook/addon-links": "^8.2.1",
+    "@storybook/addon-onboarding": "^8.2.1",
+    "@storybook/addon-styling-webpack": "^1.0.0",
+    "@storybook/nextjs": "^8.2.1",
+    "@storybook/react": "^8.2.1",
     "@testing-library/dom": "^8.12.0",
     "@testing-library/jest-dom": "^5.16.3",
     "@testing-library/react": "^13.0.0",
@@ -98,14 +116,21 @@
     "prettier": "^2.6.2",
     "prettier-plugin-tailwindcss": "^0.1.8",
     "react-docgen": "^5.4.1",
-    "storybook": "^7.6.7",
+    "storybook": "^8.2.1",
     "storybook-addon-turbo-build": "^2.0.1",
-    "storybook-dark-mode": "^3.0.3",
+    "storybook-dark-mode": "^4.0.2",
     "tailwindcss": "^3.0.24",
     "ts-jest": "^27.1.4",
-    "typescript": "4.6.3"
+    "typescript": "^5.5.3",
+    "oauth4webapi": "2.12.0"
+  },
+  "overrides": {
+    "oauth4webapi": "2.12.0"
   },
   "resolutions": {
     "**/prosemirror-model": "1.21.1"
+  },
+  "patchedDependencies": {
+    "oauth4webapi@2.12.0": "patches/oauth4webapi@2.12.0.patch"
   }
 }
