@@ -1,6 +1,6 @@
 module.exports = {
   extends: [
-    'mantine',
+
     'plugin:@next/next/recommended',
     'plugin:jest/recommended',
     'plugin:storybook/recommended',
