@@ -4,7 +4,6 @@
  *
  * THIS CODE IS AUTOMATICALLY GENERATED.
  *
- * Generated by convex@1.13.0.
  * To regenerate, run `npx convex dev`.
  * @module
  */
@@ -14,6 +13,7 @@ import type {
   FilterApi,
   FunctionReference,
 } from "convex/server";
+import type * as auth from "../auth.js";
 import type * as comment from "../comment.js";
 import type * as http from "../http.js";
 import type * as post from "../post.js";
@@ -28,6 +28,7 @@ import type * as user from "../user.js";
  * ```
  */
 declare const fullApi: ApiFromModules<{
+  auth: typeof auth;
   comment: typeof comment;
   http: typeof http;
   post: typeof post;
