@@ -1,8 +1,16 @@
 export default {
     providers: [
         {
-            domain: "https://engaged-elf-17.clerk.accounts.dev",
-            applicationID: "convex"
+            domain: process.env.CONVEX_SITE_URL,
+            applicationID: "convex"   
         }
     ]
 }
+// export default {
+//     providers: [
+//         {
+//             domain: "https://engaged-elf-17.clerk.accounts.dev",
+//             applicationID: "convex"
+//         }
+//     ]
+// }
