@@ -1,1 +1,1 @@
-if(!self.define){let e,s={};const n=(n,i)=>(n=new URL(n+".js",i).href,s[n]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=n,e.onload=s,document.head.appendChild(e)}else e=n,importScripts(n),s()})).then((()=>{let e=s[n];if(!e)throw new Error(`Module ${n} didn’t register its module`);return e})));self.define=(i,a)=>{const c=e||("document"in self?document.currentScript.src:"")||location.href;if(s[c])return;let t={};const r=e=>n(e,c),d={module:{uri:c},exports:t,require:r};s[c]=Promise.all(i.map((e=>d[e]||r(e)))).then((e=>(a(...e),t)))}}define(["./workbox-9b4d2a02"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"fb4e7e32835df55e47942e918ccbc3f2"},{url:"/_next/static/Y22HjZupxxdPGHMWrv-Nd/_buildManifest.js",revision:"6e1b341b5bfa6592d739a18bd3114840"},{url:"/_next/static/Y22HjZupxxdPGHMWrv-Nd/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/1023-d68d80e2f1466398.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2117-5f2102e855f04cf1.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2324-4bce6095be104f95.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2331-a2032e431cb35f80.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2469-3c187f21fa9482cd.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2763-e37f1230739e44a9.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/2972-d3b8bc6c52d93e2f.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/3227-87600d3b9126e10c.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/36546cfe-e0221b762cc247cb.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/3866-c48635338a9da73f.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/4059-7ef74a0ef821dc3a.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/42c68750-7d113232ce8798c0.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/4370-dc484266ab558a1b.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/4436-6c03f6be02adff45.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5104-922278be0d2c5ff4.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5279-a5a08e1182fae4a2.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5474-3e9314d921569f42.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/54a60aa6-61ab553964b59c70.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5926-a028ef04fbe0a66b.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5c29b783-def89e89642c5811.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/5f2489af-cfc48f82f6b5daba.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6171-33c3b7d5b58fd4a7.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6475-1028de7bfa404e12.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6787-70f980dfd8f2c5d5.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6879-ea632ff53db99633.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6963a3ef-a8aba1f7a4fed7e3.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/6be7e44c-520c534270457ffb.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/8677-a5abb109706faa99.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/8954-a5a803e9da1d9808.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/9b9ac224-8e2d574e12f0129e.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/9f7dcd40-0458003ba3ed3ffd.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/loading-fdd8c8ed0c27a104.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/page-4c0e77b2ea1dc1e2.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/page-6442f8d1d4e8fe03.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/configuracion/page-179878515b670905.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/crear/page-c12bf938c9dec88e.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/layout-ea101228b0b25b04.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/(main)/page-941fe63a86702728.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/_not-found/page-62946546aa67f25b.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/bienvenido/page-8da4a142b98d8bca.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/layout-446db49cf6df061c.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/app/sso-callback/page-3b29c2d62885d9fa.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/bc9e92e6-193c4a5500123401.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/ed038253-207252f7a12dd144.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/fd9d1056-0fbfdd137e948429.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/framework-4e5e198d24d9fe9d.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/main-app-e4a1b7a4b6e6ec28.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/main-bebf37042e8fe636.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/pages/_app-3c9ca398d360b709.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/pages/_error-cf5ca766ac8f493f.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-b4aed1bd9f51ac77.js",revision:"Y22HjZupxxdPGHMWrv-Nd"},{url:"/_next/static/css/07a57035855ba557.css",revision:"07a57035855ba557"},{url:"/_next/static/css/1d9502d6a32a766b.css",revision:"1d9502d6a32a766b"},{url:"/_next/static/css/38213ac1b88a21ea.css",revision:"38213ac1b88a21ea"},{url:"/_next/static/css/3b08637888b6b61c.css",revision:"3b08637888b6b61c"},{url:"/_next/static/css/54cf597dcfa2c536.css",revision:"54cf597dcfa2c536"},{url:"/_next/static/css/9a8bd61061bb1707.css",revision:"9a8bd61061bb1707"},{url:"/_next/static/css/eedf450e6f67799c.css",revision:"eedf450e6f67799c"},{url:"/_next/static/css/eeedbfa49c5cd77c.css",revision:"eeedbfa49c5cd77c"},{url:"/android-chrome-192x192.png",revision:"8c1a4ebe0c1e3cae6d185e3b667401ea"},{url:"/android-chrome-256x256.png",revision:"be854324bc83ad2421dca1d96bcc81a1"},{url:"/android-chrome-512x512.png",revision:"5cdfb78904ff19232d22210cf2cb31e8"},{url:"/apple-touch-icon-precomposed.png",revision:"df9318e1707a1eb62a6da72a2b451ef1"},{url:"/apple-touch-icon.png",revision:"0ab5a2250b08fe87b6c503c81a1066d4"},{url:"/browserconfig.xml",revision:"61bfd064535af0c276bb63b3fd579733"},{url:"/favicon-16x16.png",revision:"deedeb40d989553507d03c23523601ee"},{url:"/favicon-32x32.png",revision:"7a2154b491939ca6850a6b28b2c3bad7"},{url:"/favicon.ico",revision:"2e3f93813bf157f6335eb05284307800"},{url:"/girl.jpg",revision:"9c6ab3a75083f158877256cd0d735988"},{url:"/html_code.html",revision:"3a16e722a77b37f97c3564c145f55497"},{url:"/landscape404.svg",revision:"2f9652f7eaf5a89037580ea3c07f6887"},{url:"/logo.png",revision:"fe9bd64617ee671bd032edfdd35365f8"},{url:"/logologo.svg",revision:"b61f21fff624be7e78ec2d72dd14efae"},{url:"/manifest.json",revision:"46ad466319bfb2d3c0a574c9687eec4a"},{url:"/mstile-150x150.png",revision:"a17e09848072258afb448acfe037ea63"},{url:"/perfil.jpg",revision:"b09fc43cd92255a09a1bc579b9075191"},{url:"/porsche.jpg",revision:"debfd154f244e06be10ee1ff841919f2"},{url:"/reco.jpg",revision:"30073a6f3534f6cf6b9c65e2658191a8"},{url:"/safari-pinned-tab.svg",revision:"fd112cdbb76650768d4ce1856abddb46"},{url:"/site.webmanifest",revision:"2b80778dade1bccefe3d4057686b14a7"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:n,state:i})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
+if(!self.define){let e,s={};const c=(c,n)=>(c=new URL(c+".js",n).href,s[c]||new Promise((s=>{if("document"in self){const e=document.createElement("script");e.src=c,e.onload=s,document.head.appendChild(e)}else e=c,importScripts(c),s()})).then((()=>{let e=s[c];if(!e)throw new Error(`Module ${c} didn’t register its module`);return e})));self.define=(n,i)=>{const a=e||("document"in self?document.currentScript.src:"")||location.href;if(s[a])return;let t={};const r=e=>c(e,a),o={module:{uri:a},exports:t,require:r};s[a]=Promise.all(n.map((e=>o[e]||r(e)))).then((e=>(i(...e),t)))}}define(["./workbox-9b4d2a02"],(function(e){"use strict";importScripts(),self.skipWaiting(),e.clientsClaim(),e.precacheAndRoute([{url:"/_next/app-build-manifest.json",revision:"b818716affe61550dcea65f318807f1c"},{url:"/_next/static/3c9XOYSoIHZWPgr_-WlYL/_buildManifest.js",revision:"d636bfa358c14d3725a1e2ca2dba943b"},{url:"/_next/static/3c9XOYSoIHZWPgr_-WlYL/_ssgManifest.js",revision:"b6652df95db52feb4daf4eca35380933"},{url:"/_next/static/chunks/2117-5f2102e855f04cf1.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2236-f6817fda587446ff.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2299-b476a842edf5c0c8.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2331-a2032e431cb35f80.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2501-663321fc1204afb2.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2585-630af920534f874d.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/2972-d3b8bc6c52d93e2f.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/36546cfe-e0221b762cc247cb.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/42c68750-7d113232ce8798c0.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/4359-0adec1c32809804f.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/5104-922278be0d2c5ff4.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/5189-92c183219846e24e.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/5223-dfa0719da290753e.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/54a60aa6-9735e300cdea8596.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/5f2489af-cfc48f82f6b5daba.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6131-30819b0f45af85ca.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/619-db25c4d94937f2fa.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6238-a127ec9e95c3a176.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6437-f78de37f5e1df787.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6480-a0dda845e3d1242d.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6879-ea632ff53db99633.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6963a3ef-c89f42b52a94ba7e.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/6be7e44c-520c534270457ffb.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/7075-4dbd4c67ab4dcdfb.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/70e0d97a-02c2685acfa4b7ce.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/7399-a3bec06db6635827.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/7954-01f8746eb4f95a97.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/8076-7332c7450307b0ac.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/8252-dc34f8c25eebbf7b.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/8317-82d6af577f8932ea.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/8360-818555b11d89b9e4.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/loading-18e3bab4f3a731ce.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/%5BpostSlug%5D/page-cd818db2c71f69ce.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/%5Bauthor%5D/page-ec59ea7bc9ad5cc4.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/configuracion/page-110a3687c0629bdd.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/crear/page-c2cfb90754bc9d7e.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/layout-0e1641686c1edb46.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/(main)/page-210139abfaf66b98.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/_not-found/page-62946546aa67f25b.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/admin/redireccionar/page-35dd8091bcea033d.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/bienvenido/page-c58aff6aead08961.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/layout-33263feb50d28f98.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/r/%5Bid%5D/page-18d0570f02890fe6.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/app/sso-callback/page-3847def89984d25c.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/bc9e92e6-193c4a5500123401.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/fd9d1056-0fbfdd137e948429.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/framework-4e5e198d24d9fe9d.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/main-aab238e3b32efd56.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/main-app-e4a1b7a4b6e6ec28.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/pages/_app-3c9ca398d360b709.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/pages/_error-cf5ca766ac8f493f.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/chunks/polyfills-78c92fac7aa8fdd8.js",revision:"79330112775102f91e1010318bae2bd3"},{url:"/_next/static/chunks/webpack-37be37c8b34dc660.js",revision:"3c9XOYSoIHZWPgr_-WlYL"},{url:"/_next/static/css/07a57035855ba557.css",revision:"07a57035855ba557"},{url:"/_next/static/css/38213ac1b88a21ea.css",revision:"38213ac1b88a21ea"},{url:"/_next/static/css/382d495163e5c682.css",revision:"382d495163e5c682"},{url:"/_next/static/css/3b08637888b6b61c.css",revision:"3b08637888b6b61c"},{url:"/_next/static/css/4a28f185cc5f74f8.css",revision:"4a28f185cc5f74f8"},{url:"/_next/static/css/54cf597dcfa2c536.css",revision:"54cf597dcfa2c536"},{url:"/_next/static/css/59c11d7fa7556953.css",revision:"59c11d7fa7556953"},{url:"/_next/static/css/7c7ece6d1fa015eb.css",revision:"7c7ece6d1fa015eb"},{url:"/_next/static/css/8df49318f1de494d.css",revision:"8df49318f1de494d"},{url:"/android-chrome-192x192.png",revision:"8c1a4ebe0c1e3cae6d185e3b667401ea"},{url:"/android-chrome-256x256.png",revision:"be854324bc83ad2421dca1d96bcc81a1"},{url:"/android-chrome-512x512.png",revision:"5cdfb78904ff19232d22210cf2cb31e8"},{url:"/apple-touch-icon-precomposed.png",revision:"df9318e1707a1eb62a6da72a2b451ef1"},{url:"/apple-touch-icon.png",revision:"0ab5a2250b08fe87b6c503c81a1066d4"},{url:"/browserconfig.xml",revision:"61bfd064535af0c276bb63b3fd579733"},{url:"/favicon-16x16.png",revision:"deedeb40d989553507d03c23523601ee"},{url:"/favicon-32x32.png",revision:"7a2154b491939ca6850a6b28b2c3bad7"},{url:"/favicon.ico",revision:"2e3f93813bf157f6335eb05284307800"},{url:"/girl.jpg",revision:"9c6ab3a75083f158877256cd0d735988"},{url:"/html_code.html",revision:"3a16e722a77b37f97c3564c145f55497"},{url:"/landscape404.svg",revision:"2f9652f7eaf5a89037580ea3c07f6887"},{url:"/logo.png",revision:"fe9bd64617ee671bd032edfdd35365f8"},{url:"/logologo.svg",revision:"b61f21fff624be7e78ec2d72dd14efae"},{url:"/manifest.json",revision:"46ad466319bfb2d3c0a574c9687eec4a"},{url:"/mstile-150x150.png",revision:"a17e09848072258afb448acfe037ea63"},{url:"/perfil.jpg",revision:"b09fc43cd92255a09a1bc579b9075191"},{url:"/porsche.jpg",revision:"debfd154f244e06be10ee1ff841919f2"},{url:"/reco.jpg",revision:"30073a6f3534f6cf6b9c65e2658191a8"},{url:"/safari-pinned-tab.svg",revision:"fd112cdbb76650768d4ce1856abddb46"},{url:"/site.webmanifest",revision:"2b80778dade1bccefe3d4057686b14a7"}],{ignoreURLParametersMatching:[]}),e.cleanupOutdatedCaches(),e.registerRoute("/",new e.NetworkFirst({cacheName:"start-url",plugins:[{cacheWillUpdate:async({request:e,response:s,event:c,state:n})=>s&&"opaqueredirect"===s.type?new Response(s.body,{status:200,statusText:"OK",headers:s.headers}):s}]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:gstatic)\.com\/.*/i,new e.CacheFirst({cacheName:"google-fonts-webfonts",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:31536e3})]}),"GET"),e.registerRoute(/^https:\/\/fonts\.(?:googleapis)\.com\/.*/i,new e.StaleWhileRevalidate({cacheName:"google-fonts-stylesheets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:eot|otf|ttc|ttf|woff|woff2|font.css)$/i,new e.StaleWhileRevalidate({cacheName:"static-font-assets",plugins:[new e.ExpirationPlugin({maxEntries:4,maxAgeSeconds:604800})]}),"GET"),e.registerRoute(/\.(?:jpg|jpeg|gif|png|svg|ico|webp)$/i,new e.StaleWhileRevalidate({cacheName:"static-image-assets",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/image\?url=.+$/i,new e.StaleWhileRevalidate({cacheName:"next-image",plugins:[new e.ExpirationPlugin({maxEntries:64,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp3|wav|ogg)$/i,new e.CacheFirst({cacheName:"static-audio-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:mp4)$/i,new e.CacheFirst({cacheName:"static-video-assets",plugins:[new e.RangeRequestsPlugin,new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:js)$/i,new e.StaleWhileRevalidate({cacheName:"static-js-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:css|less)$/i,new e.StaleWhileRevalidate({cacheName:"static-style-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\/_next\/data\/.+\/.+\.json$/i,new e.StaleWhileRevalidate({cacheName:"next-data",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute(/\.(?:json|xml|csv)$/i,new e.NetworkFirst({cacheName:"static-data-assets",plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;const s=e.pathname;return!s.startsWith("/api/auth/")&&!!s.startsWith("/api/")}),new e.NetworkFirst({cacheName:"apis",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:16,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>{if(!(self.origin===e.origin))return!1;return!e.pathname.startsWith("/api/")}),new e.NetworkFirst({cacheName:"others",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:86400})]}),"GET"),e.registerRoute((({url:e})=>!(self.origin===e.origin)),new e.NetworkFirst({cacheName:"cross-origin",networkTimeoutSeconds:10,plugins:[new e.ExpirationPlugin({maxEntries:32,maxAgeSeconds:3600})]}),"GET")}));
