@@ -1,41 +1,49 @@
 "use client"
 import { uploadImage } from "@hooks/image"
 
+import { Skeleton, useMantineColorScheme } from "@mantine/core";
 import { es } from "./es"
 import { useCreateBlockNote } from "@blocknote/react";
 import { BlockNoteView } from "@blocknote/mantine";
+import "@blocknote/mantine/style.css";
 
-import { Skeleton, useMantineColorScheme } from "@mantine/core";
-import { useController } from "react-hook-form";
+import { set, useController } from "react-hook-form";
 import { MAXIMUM_MESSAGE_LENGTH, MINIMUM_MESSAGE_LENGTH } from "@lib/constants";
 
 export interface TextEditorProps {
   name: string;
   required?: boolean;
   editable?: boolean;
-  control: any
+  control: any;
+  editorState: { html: string; markdown: string, blocks: any[] } | undefined;
+  setEditorState: React.Dispatch<React.SetStateAction<{ html: string; markdown: string, blocks: any[] }>>;
 }
 
-const TextEditor = ({ editable, name, control }: TextEditorProps) => {
+const editorConfig = {
+  defaultStyles: true,
+  dictionary: es,
+  enableBlockNoteExtensions: true,
+};
+
+const TextEditor = ({ editable, name, control, setEditorState }: TextEditorProps) => {
   const {
     field,
+    formState,
     fieldState: { error },
   } = useController({
     name,
     control,
     rules: {
       required: "La descripcion es necesaria",
-      minLength: { value: MINIMUM_MESSAGE_LENGTH * 2, message: `Muy poco contenido, metele más` },
-      maxLength: { value: MAXIMUM_MESSAGE_LENGTH ** 20, message: `Nojoda te pasastes.` }
+      // minLength: { value: MINIMUM_MESSAGE_LENGTH * 2, message: `Muy poco contenido, metele más` },
+      // maxLength: { value: MAXIMUM_MESSAGE_LENGTH ** 20, message: `Nojoda te pasastes.` }
     },
   });
 
   const { colorScheme } = useMantineColorScheme()
 
   const theEditor = useCreateBlockNote({
-    defaultStyles: true,
-    dictionary: es,
-    enableBlockNoteExtensions: true,
+    ...editorConfig,
     uploadFile: uploadImage,
   });
 
@@ -45,19 +53,44 @@ const TextEditor = ({ editable, name, control }: TextEditorProps) => {
     field.onChange(html);
   };
 
+  const onFormSubmit = async () => {
+    // Converts the editor's contents from Block objects to HTML and store to state.
+    const blocks = theEditor.document;
+    console.log(blocks);
+    const html = await theEditor.blocksToHTMLLossy(theEditor.document);
+    const markdown = await theEditor.blocksToMarkdownLossy(theEditor.document);
+    setEditorState({ html, markdown, blocks });
+    field.onChange(blocks);
+  }
+
+
 
   return (
     <div className="min-h-32">
-
       <BlockNoteView
         editor={theEditor}
         editable={editable || true}
-        onChange={onDocumentChange}
-        theme={colorScheme}
+        onChange={onFormSubmit}
+        theme={colorScheme === "dark" ? "dark" : "light"}
       />
       {(error) && <span className="text-red-500">{error.message}</span>}
     </div>
   );
+}
+
+export const ViewPost = ({content}:{content:any[]}) => {
+  const editor = useCreateBlockNote({
+    ...editorConfig,
+    initialContent: content
+  });
+  const { colorScheme } = useMantineColorScheme()
+  return (
+    <BlockNoteView
+      editor={editor}
+      editable={false}
+      theme={colorScheme === "dark" ? "dark" : "light"}
+    />
+  )
 }
 
 export const EditorLoader = () => {
