@@ -1,4 +1,4 @@
-import { customAlphabet }from 'nanoid';
+import { customAlphabet } from 'nanoid';
 
 const alphabet = '012356789ABCDEFGHIJKLMOPQRSTUVWXYZabcdmnopqrstuvwxyz';
 export const nanoid = customAlphabet(alphabet, 7);
